pyproject.toml
[project]
name = "BoumWave"
version = "0.1.0"
description = "An easy static blog generator"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ruff>=0.14.1",
    "pydantic>=2.10.4",
    "jinja2>=3.1.0",
    "markdown>=3.5.0",
    "python-frontmatter>=1.1.0",
    "beautifulsoup4>=4.12.0",
]

[project.scripts]
bw = "boumwave.cli:main"

[tool.uv]
package = true
